+++ src/init.cpp
@@ -783,6 +783,15 @@ void SetupServerArgs() {
 #else
     hidden_args.emplace_back("-upnp");
 #endif
+#ifdef USE_NATPMP
+    gArgs.AddArg(
+        "-natpmp",
+        strprintf("Use NAT-PMP to map the listening port (default: %s)",
+                  DEFAULT_NATPMP ? "1 when listening and no -proxy" : "0"),
+        ArgsManager::ALLOW_BOOL, OptionsCategory::CONNECTION);
+#else
+    hidden_args.emplace_back("-natpmp");
+#endif // USE_NATPMP
     gArgs.AddArg("-whitebind=<addr>",
                  "Bind to given address and whitelist peers connecting to it. "
                  "Use [host]:port notation for IPv6",
@@ -1392,13 +1401,16 @@ void InitParameterInteraction() {
                 "%s: parameter interaction: -proxy set -> setting -listen=0\n",
                 __func__);
         }
-        // to protect privacy, do not use UPNP when a proxy is set. The user may
+        // to protect privacy, do not map ports when a proxy is set. The user may
         // still specify -listen=1 to listen locally, so don't rely on this
         // happening through -listen below.
         if (gArgs.SoftSetBoolArg("-upnp", false)) {
             LogPrintf(
                 "%s: parameter interaction: -proxy set -> setting -upnp=0\n",
                 __func__);
+        }
+        if (gArgs.SoftSetBoolArg("-natpmp", false)) {
+            LogPrintf("%s: parameter interaction: -proxy set -> setting -natpmp=0\n", __func__);
         }
         // to protect privacy, do not discover addresses by default
         if (gArgs.SoftSetBoolArg("-discover", false)) {
@@ -1415,6 +1427,9 @@ void InitParameterInteraction() {
             LogPrintf(
                 "%s: parameter interaction: -listen=0 -> setting -upnp=0\n",
                 __func__);
+        }
+        if (gArgs.SoftSetBoolArg("-natpmp", false)) {
+            LogPrintf("%s: parameter interaction: -listen=0 -> setting -natpmp=0\n", __func__);
         }
         if (gArgs.SoftSetBoolArg("-discover", false)) {
             LogPrintf(
@@ -2745,10 +2760,8 @@ bool AppInitMain(Config &config, RPCServer &rpcServer,
 
     Discover();
 
-    // Map ports with UPnP
-    if (gArgs.GetBoolArg("-upnp", DEFAULT_UPNP)) {
-        StartMapPort();
-    }
+    // Map ports with UPnP or NAT-PMP.
+    StartMapPort(gArgs.GetBoolArg("-upnp", DEFAULT_UPNP), gArgs.GetBoolArg("-natpmp", DEFAULT_NATPMP));
 
     CConnman::Options connOptions;
     connOptions.nLocalServices = nLocalServices;