+++ src/mapport.h
@@ -7,11 +7,23 @@
 
 /** -upnp default */
 #ifdef USE_UPNP
-static const bool DEFAULT_UPNP = USE_UPNP;
+static constexpr bool DEFAULT_UPNP = USE_UPNP;
 #else
-static const bool DEFAULT_UPNP = false;
-#endif
+static constexpr bool DEFAULT_UPNP = false;
+#endif // USE_UPNP
 
-void StartMapPort();
+#ifdef USE_NATPMP
+static constexpr bool DEFAULT_NATPMP = USE_NATPMP;
+#else
+static constexpr bool DEFAULT_NATPMP = false;
+#endif // USE_NATPMP
+
+enum class MapPortProtoFlag : unsigned int {
+    NONE = 0x00,
+    UPNP = 0x01,
+    NAT_PMP = 0x02,
+};
+
+void StartMapPort(bool use_upnp, bool use_natpmp);
 void InterruptMapPort();
 void StopMapPort();