+++ src/net.h
@@ -554,8 +554,8 @@ enum {
     LOCAL_IF,
     // address explicit bound to
     LOCAL_BIND,
-    // address reported by UPnP
-    LOCAL_UPNP,
+    // address reported by UPnP or NAT-PMP
+    LOCAL_MAPPED,
     // address explicitly specified (-externalip=)
     LOCAL_MANUAL,