+++ src/qt/optionsmodel.cpp
@@ -170,6 +170,13 @@ void OptionsModel::Init(bool resetSettings) {
         addOverriddenOption("-upnp");
     }
 
+    if (!settings.contains("fUseNatpmp")) {
+        settings.setValue("fUseNatpmp", DEFAULT_NATPMP);
+    }
+    if (!m_node.softSetBoolArg("-natpmp", settings.value("fUseNatpmp").toBool())) {
+        addOverriddenOption("-natpmp");
+    }
+
     if (!settings.contains("fListen")) {
         settings.setValue("fListen", DEFAULT_LISTEN);
     }
@@ -321,7 +328,13 @@ QVariant OptionsModel::data(const QModelIndex &index, int role) const {
                 return settings.value("fUseUPnP");
 #else
                 return false;
-#endif
+#endif // USE_UPNP
+            case MapPortNatpmp:
+#ifdef USE_NATPMP
+                return settings.value("fUseNatpmp");
+#else
+                return false;
+#endif // USE_NATPMP
             case MinimizeOnClose:
                 return fMinimizeOnClose;
 
@@ -412,7 +425,9 @@ bool OptionsModel::setData(const QModelIndex &index, const QVariant &value,
                 break;
             case MapPortUPnP: // core option - can be changed on-the-fly
                 settings.setValue("fUseUPnP", value.toBool());
-                m_node.mapPort(value.toBool());
+                break;
+            case MapPortNatpmp: // core option - can be changed on-the-fly
+                settings.setValue("fUseNatpmp", value.toBool());
                 break;
             case MinimizeOnClose:
                 fMinimizeOnClose = value.toBool();