+++ doc/release-notes.md
@@ -18,7 +18,9 @@ TODO
 
 ## Added functionality
 
-TODO
+- Added NAT-PMP port mapping support via [`libnatpmp`](https://miniupnp.tuxfamily.org/libnatpmp.html).
+  Use the `-natpmp` command line option to use NAT-PMP to map the listening port. If both UPnP
+  and NAT-PMP are enabled, a successful allocation from UPnP prevails over one from NAT-PMP.
 
 ## Deprecated functionality